Sweeping type changes

+++ Assets/Scripts/Game/Action/Internal/VelkarisMark.cs
@@ -7,7 +7,7 @@ namespace Game.Action.Internal
     [Il2CppSetOption(Option.NullChecks, false), Il2CppSetOption(Option.ArrayBoundsChecks, false)]
     public class VelkarisMark: Action, IInternal
     {
-        public VelkarisMark(int p, ushort f, ushort t): base(p)
+        public VelkarisMark(int p, int f, int t): base(p)
         {
             Maker = f;
             Target = t;
