+++ Assets/Scripts/Game/Effects/Buffs/Camouflage.cs
@@ -7,7 +7,7 @@ namespace Game.Effects.Buffs
     [Il2CppSetOption(Option.NullChecks, false), Il2CppSetOption(Option.ArrayBoundsChecks, false)]
     public class Camouflage: Effect, IOnMoveGenEffect
     {
-        public Camouflage(PieceLogic piece, sbyte duration = -1) : base(duration, 1, piece, "effect_camouflage")
+        public Camouflage(PieceLogic piece, int duration = -1) : base(duration, 1, piece, "effect_camouflage")
         {}
 
         public void OnCallMoveGen(PieceLogic caller, List<Action.Action> actions)
