+++ Assets/Scripts/Game/Effects/Debuffs/Broken.cs
@@ -5,7 +5,7 @@ namespace Game.Effects.Debuffs
     [Il2CppSetOption(Option.NullChecks, false), Il2CppSetOption(Option.ArrayBoundsChecks, false)]
     public class Broken: Effect, IOnApply, IOnRemove
     {
-        public Broken(sbyte duration, PieceLogic piece) : base(duration, 1, piece, "effect_broken")
+        public Broken(int duration, PieceLogic piece) : base(duration, 1, piece, "effect_broken")
         {}
 
         private Effect specialAbility;
