+++ Assets/Scripts/Game/Effects/Debuffs/Stunned.cs
@@ -5,7 +5,7 @@ namespace Game.Effects.Debuffs
     [Il2CppSetOption(Option.NullChecks, false), Il2CppSetOption(Option.ArrayBoundsChecks, false)]
     public class Stunned: Effect
     {
-        public Stunned(sbyte duration, PieceLogic piece) : base(duration, 1, piece, "effect_stunned")
+        public Stunned(int duration, PieceLogic piece) : base(duration, 1, piece, "effect_stunned")
         {}
 
         public override int GetValueForAI()
